refactor(gui): extract self-update and auth-firmware download into updater.rs
Second cut at lib.rs. Moves the five network-facing commands (fetch_url,
update_check/download/install, download_auth_firmware) together with the host
allowlist, the update-endpoint table, UpdateState/PendingUpdate, and the three
test modules that cover them (fetch allowlist, auth-firmware naming/hashing,
endpoint mapping).

Grouping these is worth more than the line count: the SSRF allowlist and the
SHA-256 verification are the security-relevant surface here, and they now sit
next to the tests that pin them instead of being spread across a 4000-line file.

sha256_hex stays in lib.rs -- the batch archive uses it too, unrelated to
updating -- and updater.rs imports it.

Command surface is unchanged: 45 registered before and after (40 lib + 5
updater), same 74 tests pass. lib.rs 4435 -> 3935.
